@@ -3,23 +3,22 @@ description: A common platform interface for the camera plugin.
 homepage: https://github.com/flutter/plugins/tree/master/packages/camera/camera_platform_interface
 # NOTE: We strongly prefer non-breaking changes, even at the expense of a
 # less-clean API. See https://flutter.dev/go/platform-interface-breaking-changes
-version: 1.5.0
+version: 2.0.0-nullsafety
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.0.5
-  plugin_platform_interface: ^1.0.1
-  cross_file: ^0.1.0
-  stream_transform: ^1.2.0
+  meta: ^1.3.0-nullsafety.6
+  plugin_platform_interface: ^1.1.0-nullsafety.2
+  cross_file: ^0.3.0-nullsafety
+  stream_transform: ^2.0.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  async: ^2.4.2
-  mockito: ^4.1.1
-  pedantic: ^1.8.0
+  async: ^2.5.0-nullsafety.3
+  pedantic: ^1.10.0-nullsafety.3
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
   flutter: ">=1.22.0"
