@@ -3,24 +3,28 @@ description: Demonstrates how to use the image_picker plugin.
 author: Flutter Team <flutter-dev@googlegroups.com>
 
 dependencies:
-  video_player: ^0.10.3
+  video_player: ^2.0.0-nullsafety.7
   flutter:
     sdk: flutter
-  flutter_plugin_android_lifecycle: ^1.0.2
+  flutter_plugin_android_lifecycle: ^2.0.0-nullsafety.2
   image_picker:
+    # When depending on this package from a real application you should use:
+    #   image_picker: ^x.y.z
+    # See https://dart.dev/tools/pub/dependencies#version-constraints
+    # The example app is bundled with the plugin so we use a path dependency on
+    # the parent directory to use the current plugin's version.
     path: ../
-  image_picker_for_web: ^0.1.0
 
 dev_dependencies:
   flutter_driver:
     sdk: flutter
   integration_test:
     path: ../../../integration_test
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.10.0"
