@@ -2,7 +2,7 @@ name: image_picker
 description: Flutter plugin for selecting images from the Android and iOS image
   library, and taking new pictures with the camera.
 homepage: https://github.com/flutter/plugins/tree/master/packages/image_picker/image_picker
-version: 0.6.7+22
+version: 0.7.0-nullsafety
 
 flutter:
   plugin:
@@ -16,19 +16,19 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_plugin_android_lifecycle: ^1.0.2
-  image_picker_platform_interface: ^1.1.0
+  flutter_plugin_android_lifecycle: ^2.0.0-nullsafety.2
+  image_picker_platform_interface: ^2.0.0-nullsafety
 
 dev_dependencies:
-  video_player: ^0.10.3
+  video_player: ^2.0.0-nullsafety.7
   flutter_test:
     sdk: flutter
   integration_test:
     path: ../../integration_test
-  mockito: ^4.1.3
+  mockito: ^5.0.0-nullsafety.7
   pedantic: ^1.8.0
   plugin_platform_interface: ^1.0.3
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.10.0"
